backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.3
colorama==0.4.6
dnspython==2.8.0
exceptiongroup==1.3.1
fastapi==0.136.1
firebase-admin==6.9.0
h11==0.16.0
idna==3.13
pydantic==2.13.3
pydantic_core==2.46.3
pymongo==4.17.0
python-multipart==0.0.26
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
web3>=6.20.0
python-dotenv>=1.0.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
google-generativeai>=0.7.0
deep-translator>=1.11.4
numpy>=1.26.0
tensorflow-cpu>=2.16.0
keras==3.10.0
tf-keras>=2.16.0
